Drop exception-based control flow
There's something smelly about using the ValidationError
to decide whether we should run in install-only or action mode.

Switch to a plain value returned by runtimes.validate
(it has a success: bool field)
and use that to decide if we're in install-only or full mode.

This gets rid of the extraneous `else throw` as well,
because the install-only mode will never throw an exception.