Refactor execution logic from main method.
Because most of the core execution logic in the
the same file as the program entrypoint, we cannot write
tests that spy on core functionality, since that would
require loading the main file. Loading the main file introduces
the side effect of running the entire Action due to the
(very necessary) top-level await.

This commit extract the core execution logic into run.ts, where
it can be tested with synthetic configs. Lastly, in this PR I refactored
the Preview command to use a more testable style. In the next commit,
I will demonstrate how this style, along with the refactor of main,
allows for unit tests which we previously could not write.