feat: comment on GitHub step summary
Using the GitHub step summary feature, the Pulumi output can be stored on the
GitHub step. This defeats the need to dive into the individual steps and find
the Pulumi output.

The input parameter comment-on-summary has the non-breaking default false. If
true, then the action will add the results of the Pulumi action to the GitHub
step summary.

Signed-off-by: Jan van den Berg <koozz@linux.com>