return error when login fails
Currently if the login fails as part of the GitHub action, we just
silently ignore that.  This can be very confusing when steps further
down the line fail, e.g. with "error: PULUMI_ACCESS_TOKEN must be set
for login during non-interactive CLI sessions", even thoug the user
has set the access token.

Instead of just continuing our merry way when the login fails, fail
the action and let the user know.