upgrade go used in workflows to latest (#1179)
We're using a very old version of Go here, which apparently can't be
installed at least through the GitHub action anymore.  We're not
supporting that version since a long time, and there really is no
reason why we shouldn't use the latest version of Golang here, so do
just that.