make `comment-on-pr-number` option work even outside of pull-request events (#1178)
We only allow the `comment-on-pr` option inside of pull requests.
This makes sense because we wouldn't be able to comment on anything
otherwise.  However with the `comment-on-pr-number` option gives us a
PR number to comment on, so it can make sense with any type of GitHub
actions event.  This used to work before #813, and it looks like
disallowing this was accidental rather than intentional.

Make this work again.