Add 3.120.0 to the list of known-bad versions (#1210)
Version 3.120.0 of Pulumi contains a P0 that means `Delete` operations
are largely broken. As a result, we'd really like to ensure it's used as
little as possible. This commit adds 3.120.0 to the list of "known-bad"
versions in the GitHub action, so that we'll force a download of a newer
version in the event that 3.120.0 is the default available.