only install new pulumi version when there's a newer one available
Users can pass 'latest' as the 'pulumi-version' configuration setting.
This currently means that they always download the latest version of
the CLI, regardless of whether that's already installed or not.

We can do better here and check if the latest version matches the one
already installed on the runner, and in that case skip the download.