add workflow for publishing immutable actions
We're switching over to publishing immutable actions from this repo.
For this we need to make sure that immutable actions are published at
each release of this repo.

With this, we can keep following our current release process, and
whenever a new release is published, this action will create a new
immutable actions package, that will then be available to users.