Fix the `output` command (#1327)
This commit fixes the `output` command to work without having to have a workDir specified to a location with a Pulumi.yaml file. It does this by avoiding calls to `LocalWorkspace.createOrSelectStack`/`selectStack` when the command is `output`. This allows the `output` command to work with any stack, even when there is not a local project available.

Additionally, when the command is `output`, don't attempt to comment on the PR; only actual pulumi operations (like the `up` command) should do that.